add test for table component
